+++ src/components/DevicesListItem/DevicesListItem.jsx
